refactor: centralize object check utility
