+++ src/browser/common.js
@@ -1,14 +1,3 @@
-// Shared utility functions
+import { isObject } from "../utils/objectUtils.js";
 
-/**
- * Checks if a value is a non-null object (but not an array).
- * @param {*} val
- * @returns {boolean}
- */
-function isNonNullNonArray(val) {
-  return val !== null && !Array.isArray(val);
-}
-
-export function isObject(val) {
-  return isNonNullNonArray(val) && typeof val === 'object';
-}
+export { isObject };
